cognitive-core/src/core/cognitive_core.py
"""
Cognitive Core - Sistema Multi-Agente Genérico.

Implementación del core principal basado en Google ADK y protocolo A2A.
"""

import logging
from typing import Dict, Any, List, Optional, Union
from datetime import datetime
from pathlib import Path

# Simplified imports - using basic ADK components that are available
try:
    from google.adk.agents.llm_agent import Agent
    from google.adk.tools import FunctionTool
    from google.adk.runners import Runner
except ImportError:
    # Fallback to basic imports if specific modules are not available
    Agent = None
    FunctionTool = None
    Runner = None

from ..a2a.agent_executor import AgentExecutor
from ..a2a.request_context import RequestContext
from ..a2a.event_queue import EventQueue
from ..shared.types import CoreConfig, AgentConfig, ProcessedData
from ..shared.exceptions import CoreError
from ..shared.utils import get_logger


class CognitiveCore(AgentExecutor):
    """
    Core principal del sistema multi-agente.
    
    Implementa el patrón AgentExecutor de A2A y utiliza Google ADK
    para la lógica interna de agentes.
    """
    
    def __init__(self, config: CoreConfig):
        """
        Inicializa el Cognitive Core.
        
        Args:
            config: Configuración del core
        """
        super().__init__()
        self.config = config
        self.logger = get_logger("CognitiveCore")
        
        # Componentes del core
        self.agent_manager = AgentManager(config)
        self.pipeline_processor = PipelineProcessor(config)
        self.orchestrator = Orchestrator(config)
        self.knowledge_synthesizer = KnowledgeSynthesizer(config)
        self.decision_engine = DecisionEngine(config)
        
        # Runner de ADK (si está disponible)
        self.runner = Runner() if Runner else None
        
        self.logger.info(f"Cognitive Core inicializado con modelo: {config.default_model}")
    
    async def execute(self, context: RequestContext, event_queue: EventQueue) -> None:
        """
        Implementación del método execute de AgentExecutor.
        
        Args:
            context: Contexto de la petición A2A
            event_queue: Cola de eventos A2A
        """
        try:
            self.logger.info(f"Procesando petición: {context.message}")
            
            # 1. Procesar input usando pipeline
            processed_input = await self.pipeline_processor.process(
                context.message, 
                context.metadata.get("input_type", "text")
            )
            
            # 2. Determinar agentes necesarios
            required_agents = await self.agent_manager.discover_agents(processed_input)
            
            # 3. Orquestar agentes usando ADK workflows
            agent_responses = await self.orchestrator.orchestrate(
                required_agents, 
                processed_input,
                context.metadata.get("orchestration_strategy", "sequential")
            )
            
            # 4. Sintetizar conocimiento
            synthesized_knowledge = await self.knowledge_synthesizer.synthesize(agent_responses)
            
            # 5. Generar decisión final
            final_decision = await self.decision_engine.decide(synthesized_knowledge)
            
            # 6. Enviar respuesta A2A
            await event_queue.put({
                "type": "response",
                "content": final_decision,
                "timestamp": datetime.now().isoformat(),
                "metadata": {
                    "agents_used": [agent.name for agent in required_agents],
                    "processing_time": self._calculate_processing_time(),
                    "confidence_score": synthesized_knowledge.get("confidence_score", 0.0)
                }
            })
            
            self.logger.info("Petición procesada exitosamente")
            
        except Exception as e:
            self.logger.error(f"Error procesando petición: {e}")
            await event_queue.put({
                "type": "error",
                "content": f"Error procesando petición: {str(e)}",
                "timestamp": datetime.now().isoformat()
            })
    
    def _calculate_processing_time(self) -> float:
        """Calcula el tiempo de procesamiento."""
        # Implementación simplificada
        return 0.0
    
    async def register_agent(self, agent_config: AgentConfig) -> None:
        """
        Registra un nuevo agente en el core.
        
        Args:
            agent_config: Configuración del agente
        """
        await self.agent_manager.register_agent(agent_config)
        self.logger.info(f"Agente registrado: {agent_config.name}")
    
    async def unregister_agent(self, agent_name: str) -> None:
        """
        Desregistra un agente del core.
        
        Args:
            agent_name: Nombre del agente
        """
        await self.agent_manager.unregister_agent(agent_name)
        self.logger.info(f"Agente desregistrado: {agent_name}")
    
    def get_agent_cards(self) -> List[Dict[str, Any]]:
        """
        Retorna las tarjetas de agentes disponibles.
        
        Returns:
            Lista de tarjetas de agentes
        """
        return self.agent_manager.get_agent_cards()
    
    def get_core_info(self) -> Dict[str, Any]:
        """
        Retorna información del core.
        
        Returns:
            Información del core
        """
        return {
            "name": self.config.name,
            "description": self.config.description,
            "model": self.config.default_model,
            "agents_count": len(self.agent_manager.agents),
            "available_agents": list(self.agent_manager.agents.keys()),
            "pipeline_capabilities": self.pipeline_processor.get_capabilities(),
            "orchestration_strategies": self.orchestrator.get_available_strategies(),
            "timestamp": datetime.now().isoformat()
        }


def create_cognitive_core(config: CoreConfig) -> CognitiveCore:
    """
    Crea una instancia del Cognitive Core.
    
    Args:
        config: Configuración del core
        
    Returns:
        Instancia del Cognitive Core
    """
    return CognitiveCore(config)
