cognitive-core/requirements.txt
# Google ADK y servicios de GCP
google-adk>=0.1.0
google-cloud-aiplatform>=1.40.0
google-cloud-logging>=3.8.0
google-cloud-trace>=1.10.0
google-cloud-storage>=2.14.0
google-cloud-documentai>=2.20.0
google-cloud-vision>=3.4.0

# Framework web y async
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiohttp>=3.9.0

# Configuración y validación
pydantic>=2.5.0
pyyaml>=6.0.0

# CLI
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Desarrollo
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Utilidades
python-dotenv>=1.0.0
asyncio-mqtt>=0.11.0